compute: refactor and clean up automatic shutdown strategies frontend ui
