+++ src/packages/frontend/compute/spend-limit.tsx
@@ -2,21 +2,8 @@
 Configuration to limit spending on a particular compute server.
 */
 
-import {
-  Alert,
-  Button,
-  Card,
-  Checkbox,
-  Flex,
-  Modal,
-  InputNumber,
-  Radio,
-  Space,
-  Spin,
-  Switch,
-} from "antd";
+import { Flex, Modal, InputNumber, Radio, Space, Spin, Switch } from "antd";
 import { useEffect, useState } from "react";
-import ShowError from "@cocalc/frontend/components/error";
 import { useServer } from "./compute-server";
 import Inline from "./inline";
 import { isEqual } from "lodash";
@@ -25,6 +12,7 @@ import {
   type SpendLimit as ISpendLimit,
   SPEND_LIMIT_DEFAULTS,
 } from "@cocalc/util/db-schema/compute-servers";
+import { AutomaticShutdownCard } from "./automatic-shutdown";
 
 export function SpendLimit({
   id,
@@ -49,74 +37,28 @@ export function SpendLimit({
     return <Spin />;
   }
 
-  console.log({ spendLimit, s: server?.configuration?.spendLimit });
-
-  const save = async () => {
-    try {
-      setSaving(true);
-      console.log(spendLimit);
-      await setServerConfiguration({
-        id,
-        configuration: {
-          spendLimit: { ...SPEND_LIMIT_DEFAULTS, ...spendLimit },
-        },
-      });
-    } catch (err) {
-      setError(`${err}`);
-    } finally {
-      setTimeout(() => setSaving(false), 1000);
-    }
-  };
-
   return (
-    <Card
-      styles={{
-        body: spendLimit.enabled ? undefined : { display: "none" },
+    <AutomaticShutdownCard
+      title="Spending Limit"
+      enabled={spendLimit.enabled}
+      setEnabled={(enabled) => setSpendLimit({ ...spendLimit, enabled })}
+      saving={saving}
+      setSaving={setSaving}
+      setError={setError}
+      error={error}
+      save={async () => {
+        await setServerConfiguration({
+          id,
+          configuration: {
+            spendLimit: { ...SPEND_LIMIT_DEFAULTS, ...spendLimit },
+          },
+        });
       }}
-      title={
-        <Flex style={{ alignItems: "center" }}>
-          <div>Spending Limit</div>
-          <div style={{ flex: 1 }} />
-          <Space>
-            <Checkbox
-              disabled={saving}
-              checked={spendLimit.enabled}
-              onChange={(e) => {
-                setSpendLimit({ ...spendLimit, enabled: e.target.checked });
-              }}
-            >
-              Enable{spendLimit.enabled ? "d" : ""}
-            </Checkbox>
-            <Button
-              type="primary"
-              disabled={
-                saving ||
-                isEqual(server.configuration?.spendLimit, spendLimit) ||
-                spendLimit.dollars == null
-              }
-              onClick={save}
-            >
-              Save {saving && <Spin style={{ marginLeft: "5px" }} />}
-            </Button>
-          </Space>
-          <div style={{ flex: 1 }} />
-          {server.configuration?.spendLimit?.enabled ? (
-            <Alert
-              style={{ marginLeft: "15px" }}
-              type="success"
-              showIcon
-              message="Spending Limit is Enabled"
-            />
-          ) : (
-            <Alert
-              style={{ marginLeft: "15px" }}
-              type="info"
-              showIcon
-              message="Spending Limit is NOT Enabled"
-            />
-          )}
-        </Flex>
+      savable={
+        !isEqual(server.configuration?.spendLimit, spendLimit) &&
+        spendLimit.dollars != null
       }
+      savedEnabled={server.configuration?.spendLimit?.enabled}
     >
       {help && (
         <div style={{ marginBottom: "15px" }}>
@@ -167,8 +109,8 @@ export function SpendLimit({
                 <InputNumber
                   style={{ width: "256px" }}
                   disabled={saving || !spendLimit.enabled}
-                  min={1}
-                  step={20}
+                  min={5}
+                  step={5}
                   addonBefore="$"
                   addonAfter="dollars"
                   placeholder="Max spend..."
@@ -185,12 +127,7 @@ export function SpendLimit({
           </>
         )}
       </Space>
-      <ShowError
-        error={error}
-        setError={setError}
-        style={{ width: "100%", marginTop: "15px" }}
-      />
-    </Card>
+    </AutomaticShutdownCard>
   );
 }
